refactor, added minimum invterval for revoked token removal

+++ backend/src/main/java/com/intive/shopme/registration/RevokedTokenService.java
@@ -3,6 +3,7 @@ package com.intive.shopme.registration;
 import com.intive.shopme.model.db.DbRevokedToken;
 import com.intive.shopme.model.rest.UserContext;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,32 +17,36 @@ public class RevokedTokenService {
 
     private final RevokedTokenRepository repository;
 
+    @Value("${jwt.minimum-revoked-tokens-remove-interval}")
+    private long revokedTokenInterval;
+    long lastTokenRemoval = 0L;
+
     @Autowired
     public RevokedTokenService(RevokedTokenRepository repository) {
         this.repository = repository;
     }
 
-    DbRevokedToken logout(DbRevokedToken dbRevokedToken) {
-        return repository.save(dbRevokedToken);
+    void logout(DbRevokedToken dbRevokedToken) {
+        removeExpiredTokens();
+        repository.save(dbRevokedToken);
     }
 
     public boolean isTokenRevoked(UserContext userContext) {
-
         UUID userId = userContext.getUserId();
         Date expirationDate = userContext.getExpirationDate();
         return repository.findOneByUserIdAndExpirationDate(userId, expirationDate) != null;
     }
 
-    public void removeOutdatedEntries() {
+    void removeExpiredTokens() {
         Date currentTime = new Date();
-        List<DbRevokedToken> revokedTokenList = repository.findAll();
-
-        for (DbRevokedToken dbRevokedToken:revokedTokenList) {
-if (dbRevokedToken.getExpirationDate().before(currentTime)){
-   repository.delete(dbRevokedToken);
-}
+        if (lastTokenRemoval + revokedTokenInterval < currentTime.toInstant().toEpochMilli()) {
+            List<DbRevokedToken> revokedTokenList = repository.findAll();
+            for (DbRevokedToken dbRevokedToken : revokedTokenList) {
+                if (dbRevokedToken.getExpirationDate().before(currentTime)) {
+                    repository.delete(dbRevokedToken);
+                }
+            }
+            lastTokenRemoval = System.currentTimeMillis();
         }
     }
 }
-
-
